Add basic logic for updating printer status
There are many parts to this:

Refactoring printer state management so that the same function that
updates the printer state also sends out the nessesary events.
Previously this had been handled in many different places in the code.

A bit of general cleanup of areas dealing with printer state

Update tempurature polling loop to get printer status and update
state, tempurature and progress as nessesary.

Populate Temp and progress getters.

Remove a lot of unneeded code.  We had a lot of copy pasta that needed
to be eaten.

AE-462 #time: 4h