Add test for comm's _update_progress function
This function updates the comm's progress variable based on the progress
the authentise api reports back.

Also this commit adds an assert_almost_equal test fixture, which does
basic type inspection and recusion for asserting equality of two
objects.  It should work for strings, floats, ints, and lists and
dictionarys consisting of any combination of thoses.  It also will
recurse into dictionarys.  The code comes from a stackoverflow answer,
and hasn't been throughly tested, so you are advised to test that the
assert works as you think it will for your given test before using it.

AE-462 #time: 45m