Add logic to claim node on printer connection
This will check if the node is already claimed by the current user
assosiated with the api_key in use on the printer, and if it is not, it
will get a claim code and assosiate that user.

Pylint was kinda hating life, so I also added in some extra disables to
make it stop complaining.

AE-1446 #time 3h